+++ src/app/inobeta-ui/ui/forms/array/array.ts
@@ -1,13 +1,14 @@
 import { IbFormControlBase } from "../controls/form-control-base";
 
 export interface IFormArrayOptions {
-  max?: number;
+  max: number;
+  addFieldLabel: string;
 }
 
 export interface IFormArray {
   key: string;
   fields: IbFormControlBase<any>[];
-  options?: IFormArrayOptions;
+  options?: Partial<IFormArrayOptions>;
   cols?: number;
   rows?: number;
   width?: string;
@@ -15,7 +16,7 @@ export interface IFormArray {
 
 export class IbFormArray implements IFormArray {
   key: string;
-  options?: IFormArrayOptions;
+  options: Partial<IFormArrayOptions> = {};
   fields: IbFormControlBase<any>[];
   cols: number;
   rows: number;
@@ -25,8 +26,10 @@ export class IbFormArray implements IFormArray {
   constructor(options: IFormArray) {
     this.key = options.key;
     this.fields = options.fields ?? [];
-    this.options = options.options ?? {
+    this.options = {
       max: Infinity,
+      addFieldLabel: "shared.ibForms.array.add",
+      ...options.options
     };
     this.cols = options.cols ?? 1;
     this.rows = options.rows ?? 1;