+++ .storybook/preview.ts
@@ -1,5 +1,7 @@
-import type { Preview } from "@storybook/angular";
+import { HttpClientModule } from "@angular/common/http";
+import { importProvidersFrom } from "@angular/core";
 import { setCompodocJson } from "@storybook/addon-docs/angular";
+import { applicationConfig, type Preview } from "@storybook/angular";
 import docJson from "../documentation.json";
 setCompodocJson(docJson);
 
@@ -12,7 +14,18 @@ const preview: Preview = {
         date: /Date$/,
       },
     },
+    docs: {
+      toc: {
+        headingSelector: 'h2, h3',
+        ignoreSelector: '#primary, #stories'
+      }
+    }
   },
+  decorators: [
+    applicationConfig({
+      providers: [importProvidersFrom(HttpClientModule)],
+    }),
+  ],
 };
 
 export default preview;