+++ src/app/inobeta-ui/ui/forms/array/array.ts
@@ -1,12 +1,16 @@
 import { IbFormControlBase } from "../controls/form-control-base";
 
 export interface IFormArrayOptions {
+  /** Max numbers of lines that an user can add to this form array */
   max: number;
+  /** Label for the button to add the first line (i18n) */
   addFieldLabel: string;
 }
 
 export interface IFormArray {
+  /** Property name for this form array. Corresponds to a FormGroup key */
   key: string;
+  /** Form control definition. These fields will be repeated for each line */
   fields: IbFormControlBase<any>[];
   options?: Partial<IFormArrayOptions>;
   cols?: number;