+++ src/app/examples/kai-table-example/kai-table-full-example.ts
@@ -1,7 +1,7 @@
 import { Component, ViewChild } from "@angular/core";
-import { MatTableDataSource } from "@angular/material/table";
 import { IbSelectionColumn } from "../../inobeta-ui/ui/kai-table/columns/selection-column";
 import { UserService } from "./users";
+import { IbKaiTableState } from "../../inobeta-ui/ui/kai-table/table.types";
 
 @Component({
   selector: "ib-kai-table-full-example",
@@ -10,6 +10,7 @@ import { UserService } from "./users";
       tableName="fullExample"
       [displayedColumns]="columns"
       [data]="data"
+      [state]="state"
     >
       <ib-table-action-group>
         <button
@@ -64,11 +65,6 @@ import { UserService } from "./users";
         padding: 30px;
         gap: 3em;
       }
-
-      ib-kai-table .ib-table {
-        --ib-table-header-cell-color: lightgrey;
-        --ib-table-header-cell-background-color: #309933;
-      }
     `,
   ],
   providers: [UserService],
@@ -77,9 +73,9 @@ export class IbKaiTableFullExamplePage {
   @ViewChild(IbSelectionColumn, { static: true })
   selectionColumn: IbSelectionColumn;
 
-  dataSource = new MatTableDataSource<any>();
   data: any[] = [];
   columns = ["name", "fruit", "number", "aDate", "subscribed"];
+  state: IbKaiTableState = "idle";
 
   constructor(private userService: UserService) {}
 
@@ -88,9 +84,13 @@ export class IbKaiTableFullExamplePage {
   }
 
   getUserOrders() {
+    this.state = "loading";
     this.userService
       .getUserOrders()
-      .subscribe((orders) => (this.data = orders));
+      .subscribe((orders) => {
+        this.data = orders
+        this.state = "idle";
+      });
   }
 
   selectionChange(data) {