+++ src/app/inobeta-ui/ui/kai-table/columns/column.ts
@@ -100,7 +100,15 @@ export class IbColumn<T> implements OnDestroy, OnInit {
    * the column's name. For example, if the column is named `id`, then the rendered value will be
    * value defined by the data's `id` property.
    */
-  @Input() dataAccessor: (data: T, name: string) => string;
+  @Input() dataAccessor: (data: T, name: string) => number | string;
+  /**
+   * Data accessor function that is used to retrieve data properties for sorting
+   */
+  @Input() sortingDataAccessor: (data: T, name: string) => number | string;
+  /**
+   * Data accessor function that is used to retrieve data properties for filtering
+   */
+  @Input() filterDataAccessor: (data: T, name: string) => number | string;
 
   /**
    * Enables sorting for the column.
@@ -170,6 +178,14 @@ export class IbColumn<T> implements OnDestroy, OnInit {
       this.dataAccessor =
         this._options.defaultDataAccessor ||
         ((data: T, name: string) => (data as any)[name]);
+    }
+
+    if (!this.sortingDataAccessor) {
+      this.sortingDataAccessor = this.dataAccessor;
+    }
+
+    if (!this.filterDataAccessor) {
+      this.filterDataAccessor = this.dataAccessor;
     }
 
     if (this._table) {