+++ src/app/inobeta-ui/ui/kai-table/columns/date-column.ts
@@ -56,5 +56,4 @@ export class IbDateColumn<T> extends IbColumn<T> {
 
   /** @ignore */
   transform = { pdf: (data) => formatDate(data, this.format, this.locale) };
-  dataAccessor = (data: T, name: string) => (data as any)[name];
 }