+++ src/app/app.component.ts
@@ -1,13 +1,12 @@
-import { Component, OnInit } from "@angular/core";
+import { Component } from "@angular/core";
 import { TranslateService } from "@ngx-translate/core";
 
 @Component({
   selector: "ib-root",
   template: `<router-outlet *ngIf="translateLoaded"></router-outlet> `,
 })
-export class AppComponent implements OnInit {
+export class AppComponent {
   translateLoaded = false;
-  ngOnInit() {}
 
   constructor(private translateService: TranslateService) {
     this.translateService.setDefaultLang("it");