+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 | Issue ID | Type | Section | Description |
 | -- | -- | -- | -- |
 |DEVK-261|feat|IbDynamicForm|Added `value` input to `ib-material-form`|
+|DEVK-264|feat|Docs|Changed compodoc for storybook|
 
 ## [17.0.1] (2024-01-10)